Finsh AVR serial drivers
git-svn-id: svn://svn.code.sf.net/p/nuttx/code/trunk@3696 42af7a65-404d-4744-a932-0658087f49c3

+++ arch/avr/src/avr/avr_internal.h
@@ -76,7 +76,7 @@ extern volatile uint8_t *current_regs;
  * heap is CONFIG_DRAM_END
  */
 
-extern uint8_t g_heapbase;
+extern uint16_t g_heapbase;
 
 #endif /* __ASSEMBLY__ */
